move style calcs to style folder..
