+++ src/dash-table/derived/filter/wrapperStyles.ts
@@ -4,8 +4,7 @@ import { memoizeOneFactory } from 'core/memoizer';
 
 import { VisibleColumns } from 'dash-table/components/Table/props';
 
-import { IConvertedStyle } from '../style';
-import { getFilterCellStyle, getFilterOpCellStyle } from '../edges';
+import { IConvertedStyle, getFilterCellStyle, getFilterOpCellStyle } from '../style';
 import { traverseMap2 } from 'core/math/matrixZipMap';
 
 const getter = (
